Resolve problem with overloaded strings in new Shake
New Shake library has even more overloaded results of command, command_
and cmd functions. This resulted in unambiguous types when part of the
result was ignored. In this case, value of stderr was discarded, while
stdin wasn't. Originally it would work, because value wrapped in
"Stderr" could be only "String", but now this wrapped value is
overloaded using CmdString type class. Since stderr is unused, it can be
ignored altogether.